Track ticks from processes

+++ kernel/defs.h
@@ -107,8 +107,8 @@ int             either_copyout(int user_dst, uint64 dst, void *src, uint64 len);
 int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
 void            procdump(void);
 void            srandom(int seed);
-uint64          random(void);
-uint64          random_number(uint64 total_tickets);
+int          random(void);
+int          random_number(uint64 total_tickets);
 
 // swtch.S
 void            swtch(struct context*, struct context*);
